Moves reply logic into conversation-continue
