+++ lib/middleware/messages/member/conversation-continue.js
