--- lib/middleware/messages/member/parse-ask-continue-answer.js
@@ -1,23 +1,0 @@
-'use strict';
-
-const helpers = require('../../../helpers');
-
-module.exports = function parseAskContinueResponse() {
-  return (req, res, next) => {
-    const asked = helpers.template.isAskContinueTemplate(req.lastOutboundTemplate);
-    if (!asked) {
-      return next();
-    }
-
-    const text = req.macro;
-    if (helpers.macro.isConfirmedCampaign(text)) {
-      return helpers.replies.confirmedContinue(req, res);
-    }
-
-    if (helpers.macro.isDeclinedCampaign(text)) {
-      return helpers.replies.declinedContinue(req, res);
-    }
-
-    return helpers.replies.invalidAskContinueResponse(req, res);
-  };
-};
